requirements.txt
# Core dependencies
aiohttp>=3.9.0
python-telegram-bot>=21.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
